pyproject.toml
[project]
name = "exp-telegrambot"
version = "0.1.0"
description = "学习和试验 Telegram 机器人"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "telethon>=1.42.0",
]
